R1CS is only passing when all constraints refer to variables using the same index.
This can only fail if a user makes non-black-box use of the snarkVM
compiler and decides to remove declared variables.

We could refactor R1CS::enforce(...) to check for each incoming constraint if its variables are consistent with declared variables, but this would be very inefficient (or require use to refactor variables into a mapping).